lambda-ecs/api/fetch-url/requirements.txt
requests==2.32.4
newspaper3k==0.2.8
lxml[html_clean]==5.2.1
